requirements.txt
# Flask & Web
flask>=3.0.0
flask-cors>=4.0.0

# Audio/Video Processing
demucs>=4.0.0
audio-denoiser>=0.1.2
noisereduce>=3.0.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0

# Utilities
python-multipart>=0.0.6
werkzeug>=3.0.0
soundfile>=0.12.0
numpy>=1.24.0
